requirements.txt
# HTTP et API
requests==2.32.5

# Database PostgreSQL
psycopg2-binary==2.9.11

# Data processing
pandas==2.3.3
numpy==1.26.4

# Environment variables
python-dotenv==1.2.1

# Timezone
pytz==2025.2

# Progress bar
tqdm==4.67.3

# PyTorch (requis pour Chronos)
torch==2.2.2

# Chronos forecasting (Amazon Research)
git+https://github.com/amazon-science/chronos-forecasting.git

# API ML
fastapi==0.129.0
uvicorn==0.40.0
pydantic==2.12.5

# ML utilities
scikit-learn==1.7.2
scipy==1.15.3
transformers==4.57.6
accelerate==1.12.0
safetensors==0.7.0
